Move DynamicCallTree to Instrumentation directory
